Split Tracing RackHandler into own gem
