+++ zipkin-gem/lib/zipkin-tracer/careless_scribe.rb
