+++ zipkin-gem/lib/zipkin-tracer/endpoint.thrift
