+++ zipkin-gem/lib/zipkin-tracer/endpoint_constants.rb
