+++ zipkin-gem/lib/zipkin-tracer/endpoint_types.rb
