+++ zipkin-gem/lib/zipkin-tracer/version.rb
@@ -11,9 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+module ZipkinTracer
+  VERSION = "0.0.1"
+end
 
-# This file is used by Rack-based servers to start the application.
-
-require ::File.expand_path('../config/environment',  __FILE__)
-use ZipkinTracer::RackHandler
-run Zipkin::Application
