+++ zipkin-web/Gemfile
@@ -6,4 +6,5 @@ gem 'thrift_client', "0.6.3"
 gem 'zookeeper', "1.2.4"
 gem 'finagle-thrift', '1.2.0'
 gem 'scribe', '0.2.4'
+gem 'zipkin-tracer', '0.0.1'
